test(meetup): use RequestBuilder for all tests
